requirements.txt
alembic==1.12.0
APScheduler==3.8.1
asgiref==3.7.2
Babel==2.8.0
blinker==1.4
certifi==2023.5.7
charset-normalizer==3.2.0
cliapp==1.0.9
click==8.0.3
# cmdtest==0.32+git
colorama==0.4.4
# command-not-found==0.3
contourpy==1.1.0
coverage==7.3.2
cryptography==3.4.8
cycler==0.11.0
# dbus-python==1.2.18
distro==1.7.0
distro-info===1.0
Django==4.2.4
django-cors-headers==4.2.0
django-versatileimagefield==3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
docutils==0.20.1
email-validator==2.0.0.post2
et-xmlfile==1.1.0
exceptiongroup==1.1.3
Flask==1.1.0
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
flask-mongoengine==0.9.1
# Flask-MySQLdb==1.0.1
# Flask-Script==1.0.0
# flask-serialize==2.1.3
# Flask-SQLAlchemy==2.5.1
# Flask-WTF==0.14.3
fonttools==4.42.0
# gpg===1.10.0
greenlet==1.1.2
httplib2==0.20.2
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
iniconfig==2.0.0
itsdangerous==2.0.1
jeepney==0.7.1
Jinja2==3.0.3
keyring==23.5.0
Kivy==2.2.1
Kivy-Garden==0.1.5
# kivymd==1.1.
# kiwisolver==1.3.1
# launchpadlib==1.10.16
# lazr.restfulclient==0.14.4
# lazr.uri==1.0.6
# libcomps==0.1.20.post1
# Mako==1.1.3
# Markdown==3.3.6
# MarkupSafe==2.0.1
# matplotlib==3.7.2
# mongoengine==0.23.1
# more-itertools==8.10.0
mysqlclient==2.1.1
# netifaces==0.11.0
# numpy==1.25.1
# oauthlib==3.2.0
# openpyxl==3.1.2
# packaging==23.1
# pandas==2.0.3
# permissive-dict==1.0.4
# Pillow==8.4.0
# pluggy==1.3.0
# pycryptodome==3.12.0
# Pygments==2.15.1
# PyGObject==3.42.1
# pyinotify==0.9.6
# PyJWT==2.3.0
# pymongo==3.12.3
# PyMySQL==1.1.0
# pyOpenSSL==21.0.0
# pyparsing==3.0.6
# pytest==7.4.2
# pytest-cov==4.1.0
# pytest-django==4.5.2
# python-apt==2.4.0+ubuntu1
# python-dateutil==2.8.2
# python-magic==0.4.27
# pytz==2021.3
# pytz-deprecation-shim==0.1.0.post0
# PyYAML==5.4.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
# rpm==4.17.0
# SecretStorage==3.3.1
# simplejson==3.17.6
# six==1.16.0
# SQLAlchemy==1.4.31
# sqlparse==0.4.4
# systemd-python==234
# tomli==2.0.1
# ttystatus==0.38
# typing_extensions==4.0.1
# tzdata==2023.3
# tzlocal==4.1
# ubuntu-advantage-tools==8001
# ufw==0.36.1
# unattended-upgrades==0.1
# urllib3==2.0.3
# wadllib==1.3.6
# Werkzeug==2.0.2
# WTForms==3.0.0
# zipp==3.6.0
